requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
numpy
streamlit
streamlit-chat
openvino-nightly
sentencepiece
gradio
auto_gptq
onnx>=1.11.0
onnxruntime
torch==2.0.1
transformers>=4.33.0
git+https://github.com/huggingface/optimum-intel.git
git+https://github.com/openvinotoolkit/nncf.git@release_v270